Ran all submissions with associated ps. Added export to code that were missing it
