+++ Hjelt/main.py
@@ -6,9 +6,9 @@ groupname="Hjelt"
 turtle.speed(0) # 0-10, 0 fastest, 10 slowest. Save precious time only running slow if you need to.
 turtle.colormode(255) # allows you to use (r, g, b) tuples as colors, with values 0-255
 turtle.hideturtle()
-
+turtle.tracer(0, 0)
 #### Code ####
-filepath = "./napoleon.jpeg"
+filepath = "Hjelt\\napoleon.jpeg"
 img = Image.open(filepath)
 resolution = (img.size[0],img.size[1])
 out_resolution = (118,161)
@@ -41,7 +41,8 @@ for y in range(out_resolution[1]):
     turtle.forward(1)
     turtle.left(90)
     turtle.pendown()
-
+    turtle.update()
+turtle.update()
 
 
 ########### export #########
