--- Mond.ran()/group.md
@@ -1,13 +1,0 @@
-**Group Name**  
-Mond.ran()
-
-**Usernames**   
-dc711   
-hg510   
-hg520   
-sw1005  
-rz326   
-
-**Title**: Mondrian Exeter  
-**Description**: A bird's eye of Streatham Campus in the style of *Piet Mondrian's* works using minimal colours and bold 
-lines
