+++ Template/YourGroupName.ps
@@ -1,7 +1,7 @@
 %!PS-Adobe-3.0 EPSF-3.0
 %%Creator: Tk Canvas Widget
 %%Title: Window .!canvas
-%%CreationDate: Mon Oct 31 12:11:50 2022
+%%CreationDate: Mon Oct 31 14:39:43 2022
 %%BoundingBox: -53 92 666 700
 %%Pages: 1
 %%DocumentData: Clean7Bit
@@ -85,74 +85,24 @@ grestore
 gsave
 grestore
 gsave
--200 501 moveto
--100 501 lineto
--200 501 lineto
--200 401 lineto
--100 401 lineto
--200 401 lineto
--200 301 lineto
--100 301 lineto
-1 setlinecap
-1 setlinejoin
-10 setlinewidth
-[] 0 setdash
-0.000 0.000 1.000 setrgbcolor AdjustColor
-stroke
 grestore
 gsave
--100 451 moveto
--0.3679204985864 332.263111316558 lineto
-1 setlinecap
-1 setlinejoin
-10 setlinewidth
-[] 0 setdash
-0.000 0.000 1.000 setrgbcolor AdjustColor
-stroke
-grestore
-gsave
--20 451 moveto
--95 321.096189432334 lineto
-1 setlinecap
-1 setlinejoin
-10 setlinewidth
-[] 0 setdash
-0.000 0.000 1.000 setrgbcolor AdjustColor
-stroke
-grestore
-gsave
-0 501 moveto
-100 501 lineto
-0 501 lineto
-5.72118872610982e-16 401 lineto
-100 401 lineto
+0 401 moveto
+-9 406 lineto
+-7 401 lineto
+-9 396 lineto
 0 401 lineto
-5.72118872610982e-16 301 lineto
-100 301 lineto
-1 setlinecap
-1 setlinejoin
-10 setlinewidth
-[] 0 setdash
-0.000 0.000 1.000 setrgbcolor AdjustColor
-stroke
-grestore
-gsave
-100 301 moveto
-91 306 lineto
-93 301 lineto
-91 296 lineto
-100 301 lineto
-0.000 0.000 1.000 setrgbcolor AdjustColor
+0.000 0.000 0.000 setrgbcolor AdjustColor
 eofill
-100 301 moveto
-91 306 lineto
-93 301 lineto
-91 296 lineto
-100 301 lineto
+0 401 moveto
+-9 406 lineto
+-7 401 lineto
+-9 396 lineto
+0 401 lineto
 1 setlinejoin 1 setlinecap
 1 setlinewidth
 [] 0 setdash
-0.000 0.000 1.000 setrgbcolor AdjustColor
+0.000 0.000 0.000 setrgbcolor AdjustColor
 stroke
 grestore
 restore showpage
