+++ WallTable/walltable.ps
@@ -1,7 +1,7 @@
 %!PS-Adobe-3.0 EPSF-3.0
 %%Creator: Tk Canvas Widget
 %%Title: Window .!canvas
-%%CreationDate: Mon Oct 31 12:11:50 2022
+%%CreationDate: Mon Oct 31 14:56:45 2022
 %%BoundingBox: -53 92 666 700
 %%Pages: 1
 %%DocumentData: Clean7Bit
@@ -83,76 +83,45 @@ save
 gsave
 grestore
 gsave
-grestore
-gsave
--200 501 moveto
--100 501 lineto
--200 501 lineto
--200 401 lineto
--100 401 lineto
--200 401 lineto
--200 301 lineto
--100 301 lineto
+0 401 moveto
+6.12323399573677e-15 501 lineto
+30 501 lineto
+39.5670858091272 499.096988312782 lineto
+47.6776695296637 493.677669529664 lineto
+53.0969883127822 485.567085809127 lineto
+55 476 lineto
+53.0969883127822 466.432914190873 lineto
+47.6776695296637 458.322330470336 lineto
+39.5670858091272 452.903011687218 lineto
+30 451 lineto
+15 451 lineto
+72.4533332339233 402.79092927351 lineto
 1 setlinecap
 1 setlinejoin
-10 setlinewidth
+1 setlinewidth
 [] 0 setdash
-0.000 0.000 1.000 setrgbcolor AdjustColor
+0.000 0.000 0.000 setrgbcolor AdjustColor
 stroke
 grestore
 gsave
--100 451 moveto
--0.3679204985864 332.263111316558 lineto
-1 setlinecap
-1 setlinejoin
-10 setlinewidth
-[] 0 setdash
-0.000 0.000 1.000 setrgbcolor AdjustColor
-stroke
 grestore
 gsave
--20 451 moveto
--95 321.096189432334 lineto
-1 setlinecap
-1 setlinejoin
-10 setlinewidth
-[] 0 setdash
-0.000 0.000 1.000 setrgbcolor AdjustColor
-stroke
-grestore
-gsave
-0 501 moveto
-100 501 lineto
-0 501 lineto
-5.72118872610982e-16 401 lineto
-100 401 lineto
-0 401 lineto
-5.72118872610982e-16 301 lineto
-100 301 lineto
-1 setlinecap
-1 setlinejoin
-10 setlinewidth
-[] 0 setdash
-0.000 0.000 1.000 setrgbcolor AdjustColor
-stroke
-grestore
-gsave
-100 301 moveto
-91 306 lineto
-93 301 lineto
-91 296 lineto
-100 301 lineto
-0.000 0.000 1.000 setrgbcolor AdjustColor
+72.4533332339233 402.79092927351 moveto
+68.7728712942852 412.406239976283 lineto
+67.0910221320905 407.290442541315 lineto
+62.3449951974198 404.745795545094 lineto
+72.4533332339233 402.79092927351 lineto
+0.000 0.000 0.000 setrgbcolor AdjustColor
 eofill
-100 301 moveto
-91 306 lineto
-93 301 lineto
-91 296 lineto
-100 301 lineto
+72.4533332339233 402.79092927351 moveto
+68.7728712942852 412.406239976283 lineto
+67.0910221320905 407.290442541315 lineto
+62.3449951974198 404.745795545094 lineto
+72.4533332339233 402.79092927351 lineto
 1 setlinejoin 1 setlinecap
 1 setlinewidth
 [] 0 setdash
-0.000 0.000 1.000 setrgbcolor AdjustColor
+0.000 0.000 0.000 setrgbcolor AdjustColor
 stroke
 grestore
 restore showpage
