+++ erhjje/bonus_submission/bonus.py
@@ -2,14 +2,14 @@ import cv2
 import turtle
 import os
 ##### Config #####
-groupname="erhjje"
+groupname="erhjjebonus_submission"
 turtle.speed(0) # 0-10, 0 fastest, 10 slowest. Save precious time only running slow if you need to.
 turtle.colormode(255) # allows you to use (r, g, b) tuples as colors, with values 0-255
 
 #### Code ####
 # Binary Image
 
-img = cv2.imread('bonus_image.png', 2)
+img = cv2.imread('erhjje\\bonus_submission\\bonus_image.png', 2)
 ret, bw_img = cv2.threshold(img, 127, 255, cv2.THRESH_BINARY)
 width = int(img.shape[1])
 height = int(img.shape[0])
