+++ winners/winners.ps
@@ -1,7 +1,7 @@
 %!PS-Adobe-3.0 EPSF-3.0
 %%Creator: Tk Canvas Widget
 %%Title: Window .!canvas
-%%CreationDate: Mon Oct 31 12:11:50 2022
+%%CreationDate: Mon Oct 31 14:49:22 2022
 %%BoundingBox: -53 92 666 700
 %%Pages: 1
 %%DocumentData: Clean7Bit
@@ -85,74 +85,99 @@ grestore
 gsave
 grestore
 gsave
--200 501 moveto
--100 501 lineto
--200 501 lineto
--200 401 lineto
--100 401 lineto
--200 401 lineto
--200 301 lineto
--100 301 lineto
+-200 451 moveto
+-100 451 lineto
+-100 481 lineto
+-160 481 lineto
+-160 541 lineto
+-100 541 lineto
+-100 571 lineto
+-160 571 lineto
+-160 631 lineto
+-100 631 lineto
+-100 661 lineto
+-200 661 lineto
+-200 451 lineto
 1 setlinecap
 1 setlinejoin
-10 setlinewidth
+1 setlinewidth
 [] 0 setdash
-0.000 0.000 1.000 setrgbcolor AdjustColor
+0.000 0.000 0.000 setrgbcolor AdjustColor
 stroke
 grestore
 gsave
--100 451 moveto
--0.3679204985864 332.263111316558 lineto
+-49.9999999999999 651 moveto
+-17.860619515673 689.302222155949 lineto
+31.3797681349374 697.984631039295 lineto
+92.0015463998481 662.984631039295 lineto
+202.001546399848 472.459042206719 lineto
+278.605990711746 408.180281238065 lineto
+334.9875479589 387.659072638525 lineto
+391.369105206055 408.180281238065 lineto
+406.369105206055 434.161043351598 lineto
 1 setlinecap
 1 setlinejoin
-10 setlinewidth
+6 setlinewidth
 [] 0 setdash
-0.000 0.000 1.000 setrgbcolor AdjustColor
+0.000 0.502 0.000 setrgbcolor AdjustColor
 stroke
 grestore
 gsave
--20 451 moveto
--95 321.096189432334 lineto
+26.3691052060549 434.161043351598 moveto
+86.3691052060549 434.161043351598 lineto
+56.3691052060549 434.161043351598 lineto
+263.201104848179 607.713697966964 lineto
+293.201104848179 607.713697966964 lineto
+233.201104848179 607.713697966964 lineto
 1 setlinecap
 1 setlinejoin
-10 setlinewidth
+6 setlinewidth
 [] 0 setdash
-0.000 0.000 1.000 setrgbcolor AdjustColor
+0.000 0.502 0.000 setrgbcolor AdjustColor
 stroke
 grestore
 gsave
-0 501 moveto
-100 501 lineto
-0 501 lineto
-5.72118872610982e-16 401 lineto
-100 401 lineto
-0 401 lineto
-5.72118872610982e-16 301 lineto
-100 301 lineto
-1 setlinecap
-1 setlinejoin
-10 setlinewidth
-[] 0 setdash
-0.000 0.000 1.000 setrgbcolor AdjustColor
-stroke
 grestore
 gsave
-100 301 moveto
-91 306 lineto
-93 301 lineto
-91 296 lineto
-100 301 lineto
-0.000 0.000 1.000 setrgbcolor AdjustColor
+233.201104848179 607.713697966964 moveto
+242.201104848179 602.713697966964 lineto
+240.201104848179 607.713697966964 lineto
+242.201104848179 612.713697966964 lineto
+233.201104848179 607.713697966964 lineto
+0.000 0.502 0.000 setrgbcolor AdjustColor
 eofill
-100 301 moveto
-91 306 lineto
-93 301 lineto
-91 296 lineto
-100 301 lineto
+233.201104848179 607.713697966964 moveto
+242.201104848179 602.713697966964 lineto
+240.201104848179 607.713697966964 lineto
+242.201104848179 612.713697966964 lineto
+233.201104848179 607.713697966964 lineto
 1 setlinejoin 1 setlinecap
 1 setlinewidth
 [] 0 setdash
-0.000 0.000 1.000 setrgbcolor AdjustColor
+0.000 0.502 0.000 setrgbcolor AdjustColor
+stroke
+grestore
+gsave
+grestore
+gsave
+grestore
+gsave
+0 401 moveto
+-9 406 lineto
+-7 401 lineto
+-9 396 lineto
+0 401 lineto
+0.000 0.000 0.000 setrgbcolor AdjustColor
+eofill
+0 401 moveto
+-9 406 lineto
+-7 401 lineto
+-9 396 lineto
+0 401 lineto
+1 setlinejoin 1 setlinecap
+1 setlinewidth
+[] 0 setdash
+0.000 0.000 0.000 setrgbcolor AdjustColor
 stroke
 grestore
 restore showpage
